// src/api/models/registerDto.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * Code Dual API
 * The Code Dual API description
 * OpenAPI spec version: 1.0
 */

export interface RegisterDto {
  /** The email of the user */
  email: string;
  /** The password of the user */
  password: string;
  /** The username of the user */
  username: string;
  /** The password of the user */
  passwordConfirmation: string;
}